fix(main):fixed GetBlobber issue on blobber registration
